+++ esp-wifi/src/ble/npl.rs
@@ -62,16 +62,10 @@ struct Event {
     queued: bool,
 }
 
-static BT_RECEIVE_QUEUE: Mutex<RefCell<Vec<ReceivedPacket>>> = Mutex::new(RefCell::new(Vec::new()));
+pub(super) static BT_RECEIVE_QUEUE: Mutex<RefCell<Vec<ReceivedPacket>>> = Mutex::new(RefCell::new(Vec::new()));
 
 #[cfg(esp32c2)]
 type OsMembufT = u32;
-
-#[derive(Debug, Clone)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct ReceivedPacket {
-    pub data: Vec<u8>,
-}
 
 /// Memory pool
 #[repr(C)]
@@ -1350,7 +1344,7 @@ unsafe extern "C" fn ble_hs_rx_data(om: *const OsMbuf, arg: *const c_void) -> i3
             data: Vec::from(&data[..data_slice.len() + 1]),
         });
 
-        dump_packet_info(&data[..(len + 1) as usize]);
+        super::dump_packet_info(&data[..(len + 1) as usize]);
     });
 
     r_os_mbuf_free_chain(om as *mut _);
@@ -1359,54 +1353,6 @@ unsafe extern "C" fn ble_hs_rx_data(om: *const OsMbuf, arg: *const c_void) -> i3
     crate::ble::controller::asynch::hci_read_data_available();
 
     0
-}
-
-static BLE_HCI_READ_DATA: Mutex<RefCell<Vec<u8>>> = Mutex::new(RefCell::new(Vec::new()));
-
-#[cfg(feature = "async")]
-pub fn have_hci_read_data() -> bool {
-    critical_section::with(|cs| {
-        let queue = BT_RECEIVE_QUEUE.borrow_ref_mut(cs);
-        !queue.is_empty() || BLE_HCI_READ_DATA.borrow_ref(cs).len() > 0
-    })
-}
-
-pub(crate) fn read_next(data: &mut [u8]) -> usize {
-    critical_section::with(|cs| {
-        let mut queue = BT_RECEIVE_QUEUE.borrow_ref_mut(cs);
-
-        match queue.pop() {
-            Some(packet) => {
-                data[..packet.data.len() as usize]
-                    .copy_from_slice(&packet.data[..packet.data.len() as usize]);
-                packet.data.len() as usize
-            }
-            None => 0,
-        }
-    })
-}
-
-pub fn read_hci(data: &mut [u8]) -> usize {
-    critical_section::with(|cs| {
-        let mut hci_read_data = BLE_HCI_READ_DATA.borrow_ref_mut(cs);
-
-        if hci_read_data.len() == 0 {
-            let mut queue = BT_RECEIVE_QUEUE.borrow_ref_mut(cs);
-
-            if let Some(packet) = queue.pop() {
-                hci_read_data.extend(packet.data);
-            }
-        }
-
-        let l = usize::min(hci_read_data.len(), data.len());
-        if l > 0 {
-            data[..l].copy_from_slice(&hci_read_data[..l]);
-            hci_read_data.drain(..l);
-            l
-        } else {
-            0
-        }
-    })
 }
 
 pub fn send_hci(data: &[u8]) {
@@ -1465,12 +1411,3 @@ pub fn send_hci(data: &[u8]) {
     }
 }
 
-#[allow(unreachable_code, unused_variables)]
-fn dump_packet_info(buffer: &[u8]) {
-    #[cfg(not(feature = "dump-packets"))]
-    return;
-
-    critical_section::with(|cs| {
-        info!("@HCIFRAME {:?}", buffer);
-    });
-}
